+++ lib/services/services.dart
@@ -1,5 +1,6 @@
 export 'src/account_service.dart';
 export 'src/api_client.dart';
+export 'src/area_service.dart';
 export 'src/authentication_service.dart';
 export 'src/business_units_service.dart';
 export 'src/departments_service.dart';