+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## 0.4.6
+
+* Add Endpoint for IslandGroup/Regional/Interal Area 
+
+## 0.4.5
+
+* Add IslandGroup/Regional/Interal Area Models and Update the depot added the InternalAreaId
+
 ## 0.4.4
 
 * Fix conflict